refactoring: Formular.jsx into small components for better sustainability and rendering of web page.
