+++ frontend/my-react-app/src/utils.js
@@ -1,0 +1,8 @@
+export const formatDate = (dateString) => {
+  if (!dateString) return '';
+  try {
+    return new Date(dateString).toLocaleString();
+  } catch {
+    return dateString;
+  }
+};
